Refactor build process, update project configs, and clean up documentation
The build process has been refactored to improve maintainability and productivity. This includes simplifying the .gitlab-ci.yml file by utilizing a more updated image which has the necessary dependencies pre-installed.

The Makefile has been heavily restructured to improve understandability and increase the flexibility of the build process. With these changes, the dependencies installation, uninstallation and removal of build artifacts can now be handled directly from the Makefile.

The configuration in project.json was updated, reducing redundancy and clarifying output paths.

In addition, README.md has been cleaned up to improve readability and keep the documentation up-to-date with the codebase changes. For instance, unnecessary functions in the usage section have been removed and links in the table of contents have been fixed.

Lastly, the copyright dates in the LICENSE file have been extended to include 2023 and the format improved for better readability.